+++ app/src/main/java/ru/drudenko/dnd/activity/MainActivity.java
@@ -50,7 +50,7 @@ public class MainActivity extends AppCompatActivity implements UpdateNotice {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         mAppBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.nav_all_spells, R.id.nav_favorite_spells, R.id.nav_all_monsters, R.id.nav_favorite_monsters, R.id.nav_master_help)
+                R.id.nav_all_spells, R.id.nav_all_monsters, R.id.nav_master_help)
                 .setDrawerLayout(drawer)
                 .build();
