--- app/src/main/java/ru/drudenko/dnd/fragment/monster/MonstersFavoriteFragment.java
@@ -1,14 +1,0 @@
-package ru.drudenko.dnd.fragment.monster;
-
-import java.util.List;
-
-import ru.drudenko.dnd.di.App;
-import ru.drudenko.dnd.model.monster.Monster;
-
-public class MonstersFavoriteFragment extends AbstractMonsterListFragment {
-
-    @Override
-    protected List<Monster> getMonstersList() {
-        return ((App) getActivity().getApplication()).monstersFavorite;
-    }
-}
