--- app/src/main/java/ru/drudenko/dnd/fragment/spell/SpellsFavoriteFragment.java
@@ -1,15 +1,0 @@
-package ru.drudenko.dnd.fragment.spell;
-
-import java.util.List;
-
-import ru.drudenko.dnd.di.App;
-import ru.drudenko.dnd.model.magic.Spell;
-
-public class SpellsFavoriteFragment extends AbstractSpellListFragment {
-
-    @Override
-    protected List<Spell> getSpellList() {
-        return ((App) getActivity().getApplication()).spellsFavorite;
-    }
-
-}
