+++ app/src/main/res/menu/search_favorite_list.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/action_search"
+        android:icon="@android:drawable/ic_menu_search"
+        android:textColor="@color/textColor"
+        android:title="@string/search_spells"
+        app:actionViewClass="androidx.appcompat.widget.SearchView"
+        app:showAsAction="ifRoom|collapseActionView" />
+
+    <item
+        android:id="@+id/action_favorite"
+        android:icon="@drawable/stars_off"
+        android:title=""
+        app:showAsAction="ifRoom" />
+</menu>
