+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -12,22 +12,10 @@
         tools:layout="@layout/fragment_spells_all" />
 
     <fragment
-        android:id="@+id/nav_favorite_spells"
-        android:name="ru.drudenko.dnd.fragment.spell.SpellsFavoriteFragment"
-        android:label="@string/menu_favorite_spells"
-        tools:layout="@layout/activity_favorite" />
-
-    <fragment
         android:id="@+id/nav_all_monsters"
-        android:name="ru.drudenko.dnd.fragment.monster.MonstersAllFragment"
+        android:name="ru.drudenko.dnd.fragment.monster.MonsterListFragment"
         android:label="@string/menu_all_monsters"
         tools:layout="@layout/fragment_spells_all" />
-
-    <fragment
-        android:id="@+id/nav_favorite_monsters"
-        android:name="ru.drudenko.dnd.fragment.monster.MonstersFavoriteFragment"
-        android:label="@string/menu_favorite_monsters"
-        tools:layout="@layout/activity_favorite" />
 
     <fragment
         android:id="@+id/nav_master_help"
